add: init
