+++ models/fts_data.go
@@ -1,0 +1,119 @@
+package models
+
+// TODO: trie
+// root depart map[region]struct{
+// Name: }map[]
+
+type TaxDepart struct {
+	Code     int
+	Name     string
+	Branches map[string]TaxDepart
+}
+
+var SupportedTaxDepartment = map[string]string{
+	"0000": "ФНС России",
+	"0100": "УФНС России по Республике Адыгея",
+	"0200": "УФНС России по Республике Башкортостан",
+	"0400": "Управление ФНС России по Республике Алтай",
+	"0500": "УФНС России по Республике Дагестан",
+	"0600": "Управление ФНС России по Республике Ингушетия",
+	"0700": "УФНС России по КБР",
+	"0800": "Управление ФНС России по Республике Калмыкия",
+	"0900": "Управление ФНС России по Карачаево-Черкесской Республике",
+	"1000": "УФНС России по Республике Карелия",
+	"1100": "УФНС России по Республике Коми",
+	"1200": "УФНС России по Республике Марий Эл",
+	"1300": "УФНС России по Республике Мордовия",
+	"1400": "Управление ФНС России по Республике Саха (Якутия)",
+	"1500": "Управление ФНС России по Республике Северная Осетия-Алания",
+	"1600": "УФНС России по Республике Татарстан",
+	"1700": "УФНС России по Республике Тыва",
+	"1800": "УФНС России по Удмуртской Республике",
+	"2000": "УФНС России по Чеченской Республике",
+	"2100": "УФНС России по Чувашской Республике",
+	"2200": "Управление ФНС России по Алтайскому краю",
+	"2300": "УФНС России по Краснодарскому краю",
+	"2400": "Управление ФНС России по Красноярскому краю",
+	"2700": "УФНС России по Хабаровскому краю",
+	"2800": "УФНС России по Амурской области",
+	"2900": "УФНС России по Архангельской области и Ненецкому автономному округу",
+	"3000": "УФНС России по Астраханской области",
+	"3100": "УФНС России по Белгородской области",
+	"3200": "УФНС России по Брянской области",
+	"3300": "УФНС России по Владимирской области",
+	"3400": "УФНС России по Волгоградской области",
+	"3500": "УФНС России по Вологодской области",
+	"3600": "УФНС России по Воронежской области",
+	"3700": "УФНС России по Ивановской области",
+	"3800": "Управление ФНС России по Иркутской области",
+	"3900": "УФНС России по Калининградской области",
+	"4000": "УФНС России по Калужской области",
+	"4100": "УФНС России по Камчатскому краю",
+	"4200": "УФНС России по Кемеровской области",
+	"4300": "УФНС России по Кировской области",
+	"4400": "УФНС России по Костромской области",
+	"4500": "УФНС России по Курганской области",
+	"4600": "УФНС России по Курской области",
+	"4700": "Управление ФНС России по Ленинградской области",
+	"4800": "Управление ФНС России по Липецкой области",
+	"4900": "УФНС России по Магаданской области",
+	"5000": "УФНС России по Московской области",
+	"5100": "УФНС России по Мурманской области",
+	"5200": "УФНС России по Нижегородской области",
+	"5300": "УФНС России по Новгородской области",
+	"5400": "УФНС России по Новосибирской области",
+	"5500": "УФНС России по Омской области",
+	"5600": "Управление ФНС России по Оренбургской области",
+	"5700": "УФНС России по Орловской области",
+	"5800": "УФНС России по Пензенской области",
+	"5900": "УФНС России по Пермскому краю",
+	"6000": "УФНС России по Псковской области",
+	"6100": "УФНС России по Ростовской области",
+	"6200": "УФНС России по Рязанской области",
+	"6300": "УФНС России по Самарской области",
+	"6400": "УФНС России по Саратовской области",
+	"6500": "УФНС России по Сахалинской области",
+	"6600": "Управление ФНС России по Свердловской области",
+	"6700": "УФНС России по Смоленской области",
+	"6800": "УФНС России по Тамбовской области",
+	"6900": "Управление ФНС России по Тверской области",
+	"7000": "УФНС России по Томской области",
+	"7100": "УФНС России по Тульской области",
+	"7200": "УФНС России по Тюменской области",
+	"7300": "УФНС России по Ульяновской области",
+	"7400": "УФНС России по Челябинской области",
+	"7500": "УФНС России по Забайкальскому краю",
+	"7600": "Управление ФНС России по Ярославской области",
+	"7700": "УФНС России по г.Москве",
+	"7900": "УФНС России по Еврейской автономной области",
+	"7800": "Управление ФНС России по Санкт-Петербургу",
+	"8600": "УФНС России по Ханты-Мансийскому автономному округу - Югре",
+	"8700": "УФНС России по Чукотскому АО",
+	"8900": "УФНС России по Ямало-Ненецкому автономному округу",
+	"9100": "УФНС России по Республике Крым",
+	"9200": "УФНС России по г.Севастополю",
+	"9800": "Организации, находящиеся в ведении ФНС России",
+	"9900": "Другие налоговые органы, подчиненные ФНС России",
+	"9951": "Межрегиональная инспекция ФНС России по Центральному федеральному округу",
+	"9952": "Межрегиональная инспекция ФНС России по Северо-Западному федеральному округу",
+	"9953": "Межрегиональная инспекция ФНС России по Южному федеральному округу",
+	"9954": "Межрегиональная инспекция ФНС России по Приволжскому федеральному округу",
+	"9955": "МИ ФНС России по Уральскому федеральному округу",
+	"9956": "Межрегиональная инспекция ФНС России по Сибирскому федеральному округу",
+	"9957": "Межрегиональная инспекция ФНС России по Дальневосточному федеральному округу",
+	"9958": "Межрегиональная инспекция ФНС России по Северо-Кавказскому федеральному округу",
+	"9959": "Межрегиональная инспекция ФНС России по Крымскому федеральному округу",
+	"9961": "МИ ФНС России по ценам",
+	"9962": "МИ ФНС России по камеральному контролю",
+	"9965": "Межрегиональная инспекция ФНС России по централизованной обработке данных",
+	"9971": "Межрегиональная инспекция ФНС России по крупнейшим налогоплательщикам № 1",
+	"9972": "МИ ФНС России по крупнейшим налогоплательщикам № 2",
+	"9973": "Межрегиональная инспекция ФНС России по крупнейшим налогоплательщикам № 3",
+	"9974": "Межрегиональная инспекция ФНС России по крупнейшим налогоплательщикам № 4",
+	"9975": "Межрегиональная инспекция ФНС России по крупнейшим налогоплательщикам № 5",
+	"9976": "Межрегиональная инспекция ФНС России по крупнейшим налогоплательщикам № 6",
+	"9977": "Межрегиональная инспекция ФНС России по крупнейшим налогоплательщикам № 7",
+	"9978": "Межрегиональная инспекция ФНС России по крупнейшим налогоплательщикам № 8",
+	"9979": "МИ ФНС России по крупнейшим налогоплательщикам № 9",
+	"9998": "ФНС России (Р)",
+}
