+++ models/models.go
@@ -1,186 +1,98 @@
 package models
 
-type SerialNumber string
+import "github.com/sshaplygin/docs-code/utils"
 
-type TaxRegionCode string
+// subjectCode
+const (
+	subjectCodeLength = 2
 
-type RegistrationReasonCode string
-
-type TaxOffice struct {
-	RegionCode TaxRegionCode
-	OfficeName string
-}
-
-type TaxDepartment struct {
-	RegionCode TaxRegionCode
-	DepartName string
-}
-
-type ReasonRegistration struct {
-	Code        RegistrationReasonCode
-	Description string
-}
-
-var (
-	SupportedTaxOfficesSet = map[TaxRegionCode]string{
-		"": "",
-	}
-
-	SupportedTaxDepartmensSet = map[TaxRegionCode]string{
-		"0000": "ФНС России",
-		"0100": "УФНС России по Республике Адыгея",
-		"0200": "УФНС России по Республике Башкортостан",
-		"0400": "Управление ФНС России по Республике Алтай",
-		"0500": "УФНС России по Республике Дагестан",
-		"0600": "Управление ФНС России по Республике Ингушетия",
-		"0700": "УФНС России по КБР",
-		"0800": "Управление ФНС России по Республике Калмыкия",
-		"0900": "Управление ФНС России по Карачаево-Черкесской Республике",
-		"1000": "УФНС России по Республике Карелия",
-		"1100": "УФНС России по Республике Коми",
-		"1200": "УФНС России по Республике Марий Эл",
-		"1300": "УФНС России по Республике Мордовия",
-		"1400": "Управление ФНС России по Республике Саха (Якутия)",
-		"1500": "Управление ФНС России по Республике Северная Осетия-Алания",
-		"1600": "УФНС России по Республике Татарстан",
-		"1700": "УФНС России по Республике Тыва",
-		"1800": "УФНС России по Удмуртской Республике",
-		"2000": "УФНС России по Чеченской Республике",
-		"2100": "УФНС России по Чувашской Республике",
-		"2200": "Управление ФНС России по Алтайскому краю",
-		"2300": "УФНС России по Краснодарскому краю",
-		"2400": "Управление ФНС России по Красноярскому краю",
-		"2700": "УФНС России по Хабаровскому краю",
-		"2800": "УФНС России по Амурской области",
-		"2900": "УФНС России по Архангельской области и Ненецкому автономному округу",
-		"3000": "УФНС России по Астраханской области",
-		"3100": "УФНС России по Белгородской области",
-		"3200": "УФНС России по Брянской области",
-		"3300": "УФНС России по Владимирской области",
-		"3400": "УФНС России по Волгоградской области",
-		"3500": "УФНС России по Вологодской области",
-		"3600": "УФНС России по Воронежской области",
-		"3700": "УФНС России по Ивановской области",
-		"3800": "Управление ФНС России по Иркутской области",
-		"3900": "УФНС России по Калининградской области",
-		"4000": "УФНС России по Калужской области",
-		"4100": "УФНС России по Камчатскому краю",
-		"4200": "УФНС России по Кемеровской области",
-		"4300": "УФНС России по Кировской области",
-		"4400": "УФНС России по Костромской области",
-		"4500": "УФНС России по Курганской области",
-		"4600": "УФНС России по Курской области",
-		"4700": "Управление ФНС России по Ленинградской области",
-		"4800": "Управление ФНС России по Липецкой области",
-		"4900": "УФНС России по Магаданской области",
-		"5000": "УФНС России по Московской области",
-		"5100": "УФНС России по Мурманской области",
-		"5200": "УФНС России по Нижегородской области",
-		"5300": "УФНС России по Новгородской области",
-		"5400": "УФНС России по Новосибирской области",
-		"5500": "УФНС России по Омской области",
-		"5600": "Управление ФНС России по Оренбургской области",
-		"5700": "УФНС России по Орловской области",
-		"5800": "УФНС России по Пензенской области",
-		"5900": "УФНС России по Пермскому краю",
-		"6000": "УФНС России по Псковской области",
-		"6100": "УФНС России по Ростовской области",
-		"6200": "УФНС России по Рязанской области",
-		"6300": "УФНС России по Самарской области",
-		"6400": "УФНС России по Саратовской области",
-		"6500": "УФНС России по Сахалинской области",
-		"6600": "Управление ФНС России по Свердловской области",
-		"6700": "УФНС России по Смоленской области",
-		"6800": "УФНС России по Тамбовской области",
-		"6900": "Управление ФНС России по Тверской области",
-		"7000": "УФНС России по Томской области",
-		"7100": "УФНС России по Тульской области",
-		"7200": "УФНС России по Тюменской области",
-		"7300": "УФНС России по Ульяновской области",
-		"7400": "УФНС России по Челябинской области",
-		"7500": "УФНС России по Забайкальскому краю",
-		"7600": "Управление ФНС России по Ярославской области",
-		"7700": "УФНС России по г.Москве",
-		"7900": "УФНС России по Еврейской автономной области",
-		"7800": "Управление ФНС России по Санкт-Петербургу",
-		"8600": "УФНС России по Ханты-Мансийскому автономному округу - Югре",
-		"8700": "УФНС России по Чукотскому АО",
-		"8900": "УФНС России по Ямало-Ненецкому автономному округу",
-		"9100": "УФНС России по Республике Крым",
-		"9200": "УФНС России по г.Севастополю",
-		"9800": "Организации, находящиеся в ведении ФНС России",
-		"9900": "Другие налоговые органы, подчиненные ФНС России",
-		"9951": "Межрегиональная инспекция ФНС России по Центральному федеральному округу",
-		"9952": "Межрегиональная инспекция ФНС России по Северо-Западному федеральному округу",
-		"9953": "Межрегиональная инспекция ФНС России по Южному федеральному округу",
-		"9954": "Межрегиональная инспекция ФНС России по Приволжскому федеральному округу",
-		"9955": "МИ ФНС России по Уральскому федеральному округу",
-		"9956": "Межрегиональная инспекция ФНС России по Сибирскому федеральному округу",
-		"9957": "Межрегиональная инспекция ФНС России по Дальневосточному федеральному округу",
-		"9958": "Межрегиональная инспекция ФНС России по Северо-Кавказскому федеральному округу",
-		"9959": "Межрегиональная инспекция ФНС России по Крымскому федеральному округу",
-		"9961": "МИ ФНС России по ценам",
-		"9962": "МИ ФНС России по камеральному контролю",
-		"9965": "Межрегиональная инспекция ФНС России по централизованной обработке данных",
-		"9971": "Межрегиональная инспекция ФНС России по крупнейшим налогоплательщикам № 1",
-		"9972": "МИ ФНС России по крупнейшим налогоплательщикам № 2",
-		"9973": "Межрегиональная инспекция ФНС России по крупнейшим налогоплательщикам № 3",
-		"9974": "Межрегиональная инспекция ФНС России по крупнейшим налогоплательщикам № 4",
-		"9975": "Межрегиональная инспекция ФНС России по крупнейшим налогоплательщикам № 5",
-		"9976": "Межрегиональная инспекция ФНС России по крупнейшим налогоплательщикам № 6",
-		"9977": "Межрегиональная инспекция ФНС России по крупнейшим налогоплательщикам № 7",
-		"9978": "Межрегиональная инспекция ФНС России по крупнейшим налогоплательщикам № 8",
-		"9979": "МИ ФНС России по крупнейшим налогоплательщикам № 9",
-		"9998": "ФНС России (Р)",
-	}
-
-	SupportedRegistrationReasonSet = map[RegistrationReasonCode]string{
-		"01": "Постановка на учет в налоговом органе российской организации в качестве налогоплательщика по месту ее нахождения",
-		"02": "Постановка на учет налогоплательщика - российской организации по месту нахождения ее филиала, исполняющего обязанности организации по уплате налогов и сборов",
-		"03": "Постановка на учет налогоплательщика - российской организации по месту нахождения ее филиала, не исполняющего обязанности организации по уплате налогов и сборов",
-		"04": "Постановка на учет налогоплательщика - российской организации по месту нахождения ее представительства, исполняющего обязанности организации по уплате налогов и сборов",
-		"05": "Постановка на учет налогоплательщика - российской организации по месту нахождения ее представительства, не исполняющего обязанности организации по уплате налогов и сборов",
-		"06": "Постановка на учет налогоплательщика - российской организации по месту нахождения принадлежащего ему недвижимого имущества (за исключением земельных участков и транспортных средств)",
-		"07": "Постановка на учет налогоплательщика - российской организации по месту нахождения принадлежащего ему недвижимого имущества (за исключением земельных участков и транспортных средств), имеющей на территории одного налогового органа число объектов имущества, превышающее 999",
-		"08": "Постановка на учет налогоплательщика - российской организации по месту нахождения принадлежащего ему земельного участка",
-		"10": "Постановка на учет налогоплательщика - российской организации по месту нахождения принадлежащих ему морских транспортных средств",
-		"12": "Постановка на учет налогоплательщика - российской организации по месту нахождения принадлежащих ему речных транспортных средств",
-		"14": "Постановка на учет налогоплательщика - российской организации по месту нахождения принадлежащих ему воздушных транспортных средств",
-		"16": "Постановка на учет налогоплательщика - российской организации по месту нахождения принадлежащих ему автомобильных транспортных средств",
-		"17": "Постановка на учет налогоплательщика - российской организации по месту нахождения принадлежащих ему автомобильных транспортных средств, число которых на территории одного налогового органа у данной организации превышает 999",
-		"18": "Постановка на учет налогоплательщика - российской организации по месту нахождения принадлежащих ему автомобильных транспортных средств, число которых на территории одного налогового органа у данной организации превышает 1998",
-		"19": "Постановка на учет налогоплательщика - российской организации по месту нахождения принадлежащих ему автомобильных транспортных средств, число которых на территории одного налогового органа у данной организации превышает 2997",
-		"26": "Постановка на учет налогоплательщика - российской организации по месту нахождения принадлежащих ему иных видов транспортных средств",
-		"27": "Постановка на учет налогоплательщика - российской организации по месту нахождения принадлежащего ему недвижимого имущества (за исключением земельных участков и транспортных средств), имеющей на территории одного налогового органа число объектов имущества, превышающее 1998",
-		"28": "Постановка на учет налогоплательщика - российской организации по месту нахождения принадлежащего ему недвижимого имущества (за исключением земельных участков и транспортных средств), имеющей на территории одного налогового органа число объектов имущества, превышающее 2997",
-		"29": "Постановка на учет налогоплательщика - российской организации по месту нахождения принадлежащих ему автомобильных транспортных средств, число которых на территории одного налогового органа у данной организации превышает 3996",
-		"30": "Постановка на учет российской организации - налогового агента, не Учтенной в качестве налогоплательщика (по месту ее нахождения)",
-		"31": "Постановка на учет налогоплательщика - российской организации по месту нахождения обособленного подразделения, в отношении которого не проведена процедура оформления в соответствии с пунктом 3 статьи 55 Гражданского кодекса Российской Федерации, исполняющего обязанности организации по уплате налогов и сборов",
-		"32": "Постановка на учет налогоплательщика - российской организации по месту нахождения обособленного подразделения, в отношении которого не проведена процедура оформления в соответствии с пунктом 3 статьи 55 Гражданского кодекса Российской Федерации, не исполняющего обязанности организации по уплате налогов и сборов",
-		"33": "Постановка на учет в налоговом органе организации по месту нахождения участка недр в качестве налогоплательщика налога на добычу полезных ископаемых",
-		"34": "Учет в налоговом органе организации в качестве налогоплательщика налога на добычу полезных ископаемых на основании сообщений от органов, осуществляющих учет и (или) регистрацию пользователей природными ресурсами, а также лицензирование деятельности, связанной с пользованием этими ресурсами, в случаях, предусмотренных пунктами приказа МНС России",
-		"35": "Постановка на учет организации - налогоплательщика единого налога на вмененный доход в налоговом органе по месту осуществления деятельности, подлежащей налогообложению единым налогом на вмененный доход.",
-		"36": "Постановка на учет в налоговом органе организации при выполнении соглашения о разделе продукции",
-		"37": "Постановка (восстановление с ранее присвоенным ИНН) на учет организации, ранее состоявшей на учете, в связи с внесением в ЕГРЮЛ записи об отмене судом решения о прекращении (ликвидации) юридического лица",
-		"40": "Учет в налоговом органе организации при возникновении у нее обязанности по уплате отдельных видов налогов, предусмотренных действующим законодательством о налогах и сборах, в случае, когда организация осуществляет деятельность вне места своего нахождения сроком менее 1 месяца, не создавая при этом обособленное подразделение и не имея недвижимое имущество или транспортные средства по месту осуществления деятельности",
-		"41": "Учет в налоговом органе организации, уплачивающей отдельные виды налогов, в случае отсутствия у нее на территории этого налогового органа оснований для постановки на учет",
-		"42": "Учет сведений об исполнении организацией - участником договора простого товарищества (договора о совместной деятельности) обязанностей по исчислению и уплате всей суммы акциза, исчисленной по операциям, осуществляемым в рамках договора простого товарищества (договора о совместной деятельности) в налоговом органе по месту ее нахождения",
-		"49": "Постановка на учет организации в качестве налогоплательщика налога на игорный бизнес в налоговом органе по месту установки объекта налогообложения",
-		"50": "Постановка на учет в налоговом органе российской организации в качестве крупнейшего налогоплательщика",
-		"51": "Постановка на учет отделения иностранной организации",
-		"52": "Постановка на учет отделения иностранной организации в Российской Федерации, созданной филиалом этой иностранной организации в иностранном государстве",
-		"53": "Постановка на учет отделения иностранной организации, созданной филиалом этой иностранной организации в Российской Федерации",
-		"60": "Постановка на учет посольства иностранного государства",
-		"61": "Постановка на учет консульства иностранного государства",
-		"62": "Постановка на учет представительства иностранного государства, приравненного к дипломатическому",
-		"63": "Постановка на учет международной организации",
-		"70": "Постановка на учет иностранной (международной) организации, имеющей недвижимое имущество в Российской Федерации, за исключением транспортных средств, относящихся к недвижимому имуществу",
-		"71": "Постановка на учет иностранной (международной) организации, имеющей транспортные средства в Российской Федерации, не относящиеся к недвижимому имуществу",
-		"72": "Постановка на учет иностранной (международной) организации, имеющей морские транспортные средства в Российской Федерации",
-		"73": "Постановка на учет иностранной (международной) организации, имеющей недвижимое имущество в Российской Федерации, за исключением транспортных средств, относящихся к недвижимому имуществу",
-		"74": "Постановка на учет иностранной (международной) организации, имеющей воздушные транспортные средства в Российской Федерации",
-		"75": "Постановка на учет иностранной (международной) организации, имеющей космические объекты в Российской Федерации",
-		"76": "Постановка на учет иностранной организации, выступающей в качестве инвестора по соглашению о разделе продукции или оператора соглашения",
-		"87": "Cчет иностранной (международной) организации в связи с открытием счетов в банках на территории Российской Федерации",
-	}
+	minSubjectCodeLength = 0
+	maxSubjectCodeLength = 99
 )
+
+// ConstitutionSubjectCode required length 2.
+// Код субъекта Российской Федерации согласно 65 статье Конституции
+type ConstitutionSubjectCode int
+
+func (csc ConstitutionSubjectCode) IsValid() bool {
+	if csc < minSubjectCodeLength || csc > maxSubjectCodeLength {
+		return false
+	}
+
+	_, ok := supportedSubjectsCodes[csc]
+	return ok
+}
+
+func (csc ConstitutionSubjectCode) String() string {
+	_, ok := supportedCountryCodes[cc]
+	if !ok {
+		return RussiaCountryCode.String()
+	}
+
+	return utils.StrCode(int(cc), countryCodeLength)
+}
+
+func (csc ConstitutionSubjectCode) GetName() string {
+	codeName, ok := supportedCountryCodes[cc]
+	if !ok {
+		return unspecifiedCountryCode
+	}
+
+	return codeName
+}
+
+type RegionTaxServiceNumber int
+
+func (rtn RegionTaxServiceNumber) IsValid() bool {
+	if csc < minSubjectCodeLength || csc > maxSubjectCodeLength {
+		return false
+	}
+
+	_, ok := supportedSubjectsCodes[csc]
+	return ok
+}
+
+func (rtn RegionTaxServiceNumber) String() string {
+	_, ok := supportedCountryCodes[cc]
+	if !ok {
+		return RussiaCountryCode.String()
+	}
+
+	return utils.StrCode(int(cc), countryCodeLength)
+}
+
+const rootTaxServiceDepart = "0000"
+
+type TaxRegionCode struct {
+	subjectCode   ConstitutionSubjectCode
+	serviceNumber RegionTaxServiceNumber
+}
+
+func (trc *TaxRegionCode) IsValid() bool {
+	if trc == nil {
+		return false
+	}
+
+	return trc.subjectCode.IsValid() && trc.serviceNumber.IsValid()
+}
+
+func (trc *TaxRegionCode) String() string {
+	if trc == nil {
+		return rootTaxServiceDepart
+	}
+
+	return trc.subjectCode.String() + trc.serviceNumber.String()
+}
+
+func (trc *TaxRegionCode) GetName() string {
+	if trc == nil {
+		return SupportedTaxDepartment[rootTaxServiceDepart]
+	}
+
+	codeName, ok := SupportedTaxDepartment[trc.String()]
+	if !ok {
+		return SupportedTaxDepartment[rootTaxServiceDepart]
+	}
+
+	return codeName
+}
