+++ models/subjects_data.go
@@ -1,0 +1,3 @@
+package models
+
+var SupportedSubjectsCodes = map[ConstitutionSubjectCode]string{}
