+++ parser/subjects/go.mod
@@ -1,0 +1,9 @@
+module github.com/sshaplygin/docs-code/parser/subjects
+
+go 1.21.5
+
+require (
+	github.com/PuerkitoBio/goquery v1.8.1 // indirect
+	github.com/andybalholm/cascadia v1.3.1 // indirect
+	golang.org/x/net v0.7.0 // indirect
+)
